refactor: update 쿼리 대신 dirtyChecking 사용
